fix(db): validate product availability during order
The previous version of this code failed to determine whether the
product is in stock before confirming it as part of the order. Stock
validation is done at other stages like when adding to cart and when
updating item quantity, but never when ordering. This will fix
a scenario where by an item with 0 quantity and 0 amount is added to the
order.